requirements.txt
requests==2.9.1
python-telegram-bot==4.0.3
lxml==3.6.0